change links common code to match rails patterns
